+++ tests/__init__.py
@@ -1,1 +1,6 @@
 """Tests."""
+
+import logging
+
+# Disable logging during tests
+logging.disable(logging.CRITICAL)