+++ tests/test_sitemap.py
@@ -5,13 +5,15 @@ Tests for the XML sitemap views.
 from flask import url_for
 from lxml import etree
 
+from kerko.index import open_index
 from kerko.searcher import Searcher
-from kerko.storage import open_index
-from tests.integration_testing import SynchronizedTestCase
+from tests.base import SyncIndexTestCase
 
 
-class PopulatedSitemapTestCase(SynchronizedTestCase):
-    """Test the sitemap with a populated library."""
+class SitemapTestCase(SyncIndexTestCase):
+    """Test the sitemap with a dummy library."""
+
+    fixture_name = "dummy"
 
     def setUp(self):
         self.namespaces = {
@@ -22,7 +24,7 @@ class PopulatedSitemapTestCase(SynchronizedTestCase):
         """Test the XML sitemap index."""
 
         with self.app.test_client() as client:
-            response = client.get(f"{self.URL_PREFIX}/sitemap.xml")
+            response = client.get(f"{self.url_prefix}/sitemap.xml")
             self.assertEqual(response.status_code, 200)
 
             root = etree.fromstring(response.get_data(as_text=True).encode("utf-8"))
@@ -35,7 +37,7 @@ class PopulatedSitemapTestCase(SynchronizedTestCase):
         """Test the XML sitemap."""
 
         with self.app.test_client() as client:
-            response = client.get(f"{self.URL_PREFIX}/sitemap1.xml")
+            response = client.get(f"{self.url_prefix}/sitemap1.xml")
             self.assertEqual(response.status_code, 200)
 
             root = etree.fromstring(response.get_data(as_text=True).encode("utf-8"))
@@ -47,7 +49,7 @@ class PopulatedSitemapTestCase(SynchronizedTestCase):
             )
             self.assertEqual(
                 len(urls),
-                len(Searcher(open_index("index")).search()),
+                len(Searcher(open_index()).search()),
                 "The number of URLs in the sitemap does not match the number of items.",
             )
             for url in urls:
@@ -57,13 +59,13 @@ class PopulatedSitemapTestCase(SynchronizedTestCase):
         """Test out of range sitemaps."""
 
         with self.app.test_client() as client:
-            response = client.get(f"{self.URL_PREFIX}/sitemap0.xml")
+            response = client.get(f"{self.url_prefix}/sitemap0.xml")
             self.assertEqual(response.status_code, 404)
 
         with self.app.test_client() as client:
-            response = client.get(f"{self.URL_PREFIX}/sitemap500.xml")
+            response = client.get(f"{self.url_prefix}/sitemap500.xml")
             self.assertEqual(response.status_code, 404)
 
         with self.app.test_client() as client:
-            response = client.get(f"{self.URL_PREFIX}/sitemapbomb.xml")
+            response = client.get(f"{self.url_prefix}/sitemapbomb.xml")
             self.assertEqual(response.status_code, 404)